requirements.txt
flask
tensorflow
pillow
numpy
gdown
gunicorn

